working on extracting the various components
