+++ public/index.html
@@ -20,6 +20,7 @@
   <style>
     span.v {
       margin: 5px;
+      color: brown
     }
   </style>
   <noscript>You need to enable JavaScript to run this app.</noscript>
